Control.Concurrent.Class.MonadSTM.Strict.TVar.Checked

Description

This module corresponds to Control.Concurrent.STM.TVar in the stm package.

This module can be used as a drop-in replacement for Control.Concurrent.Class.MonadSTM.Strict.TVar, but not the other way around.

StrictTVar

type LazyTVar m = LazyTVar m

data StrictTVar m a

castStrictTVar ∷ LazyTVar m ~ LazyTVar n ⇒ StrictTVar m a → StrictTVar n a

fromLazyTVar ∷ LazyTVar m a → StrictTVar m a

Create a StrictMVar from a LazyMVar

It is not guaranteed that the LazyTVar contains a value that is in WHNF, so there is no guarantee that the resulting StrictTVar contains a value that is in WHNF. This should be used with caution.

The resulting StrictTVar has a trivial invariant.

modifyTVar ∷ MonadSTM m ⇒ StrictTVar m a → (a → a) → STM m ()

newTVar ∷ MonadSTM m ⇒ a → STM m (StrictTVar m a)

newTVarIO ∷ MonadSTM m ⇒ a → m (StrictTVar m a)

newTVarWithInvariant ∷ (MonadSTM m, HasCallStack) ⇒ (a → Maybe String) → a → STM m (StrictTVar m a)

newTVarWithInvariantIO ∷ (MonadSTM m, HasCallStack) ⇒ (a → Maybe String) → a → m (StrictTVar m a)

readTVar ∷ MonadSTM m ⇒ StrictTVar m a → STM m a

readTVarIO ∷ MonadSTM m ⇒ StrictTVar m a → m a

stateTVar ∷ MonadSTM m ⇒ StrictTVar m s → (s → (a, s)) → STM m a

swapTVar ∷ MonadSTM m ⇒ StrictTVar m a → a → STM m a

toLazyTVar ∷ StrictTVar m a → LazyTVar m a

Get the underlying TVar

Since we obviously cannot guarantee that updates to this LazyTVar will be strict, this should be used with caution.

Similarly, we can not guarantee that updates to this LazyTVar do not break the original invariant that the StrictTVar held.

unsafeToUncheckedStrictTVar ∷ StrictTVar m a → StrictTVar m a

Create an unchecked reference to the given checked StrictTVar.

Note that the invariant is only guaranteed when modifying the checked TVar. Any modification to the unchecked reference might break the invariants.

writeTVar ∷ (MonadSTM m, HasCallStack) ⇒ StrictTVar m a → a → STM m ()

MonadLabelSTM

labelTVar ∷ MonadLabelledSTM m ⇒ StrictTVar m a → String → STM m ()

labelTVarIO ∷ MonadLabelledSTM m ⇒ StrictTVar m a → String → m ()

MonadTraceSTM

traceTVar ∷ MonadTraceSTM m ⇒ proxy m → StrictTVar m a → (Maybe a → a → InspectMonad m TraceValue) → STM m ()

traceTVarIO ∷ MonadTraceSTM m ⇒ StrictTVar m a → (Maybe a → a → InspectMonad m TraceValue) → m ()

Invariant

checkInvariant ∷ HasCallStack ⇒ Maybe String → a → a

Check invariant (if enabled) before continuing

checkInvariant mErr x is equal to x if mErr == Nothing, and throws an error err if mErr == Just err.

This is exported so that other code that wants to conditionally check invariants can reuse the same logic, rather than having to introduce new per-package flags.